# doc/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Swagger Movies",
    "description": "Please use account: `admin`, password: `123@qweQ!` to login and get token. Please config environment in the file `.env` before migrate or seed db.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "tin.quachtrung@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:{port}/api",
      "variables": {
        "port": {
          "enum": [
            "6969",
            "8082"
          ],
          "default": "8082"
        }
      }
    }
  ],
...
  "paths": {
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Login",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "requestBody": {
          "description": "Login",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true,
          "x-codegen-request-body-name": "body"
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Register",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "requestBody": {
          "description": "Register",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true,
          "x-codegen-request-body-name": "body"
        }
      }
    },
    "/user/update": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update Infomation User",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "requestBody": {
          "description": "Update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true,
          "x-codegen-request-body-name": "body"
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/user/infoUser": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Register",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "The user Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/manageuser": {
      "get": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Get list User",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "The user Name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "The group Code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageuser/role": {
      "get": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Get list role",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageuser/paging": {
      "get": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Get list User Page",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "The user Name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "The group Code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page of list user",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of list user",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageuser/search": {
      "get": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Get list User by searching",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "The user Name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "The group Code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageuser/searchpaging": {
      "get": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Get list User by Seaching page",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "The user Name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupCode",
            "in": "query",
            "description": "The group Code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page of list user",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of list user",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageuser/delete": {
      "delete": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Get list User by Seaching page",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "The user Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageuser/create": {
      "post": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Create User (role is admin)",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "requestBody": {
          "description": "Register",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true,
          "x-codegen-request-body-name": "body"
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/manageuser/update": {
      "put": {
        "tags": [
          "Mangement User"
        ],
        "summary": "Update User (role is admin)",
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "requestBody": {
          "description": "Register",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true,
          "x-codegen-request-body-name": "body"
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "type": "number"
          },
          "fullName": {
            "type": "string"
          },
          "roleCode": {
            "type": "string"
          },
          "groupCode": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}